+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * Removed "style-disabled" property and GetDisabledStyle function
 * Added GoogleFonts field to AppParams
 * Added functions: GetWhiteSpace, GetWordBreak, ScrollIntoViewIfNeeded
-* Added Popups, PopupDefault, PopupDefaultsSeq, and SetPopupDefaults methods to Session interface
+* Added ClientSession interface
+* Added ClientSession, Popups, PopupDefault, PopupDefaultsSeq, and SetPopupDefaults methods to Session interface
+* Removed ClientItem, SetClientItem, RemoveClientItem, and RemoveAllClientItems methods from Session interface
 * Added DismissWithoutAnimation add SetHotKey methods to Popup interface
 * Added "outside-color" add "outside-filter" properties to Popup interface
 * Added ViewSeq add ViewCount methods to ParentView interface